Avoid layer breaking between the chat parser and dynamic file references
