+++ src/vs/workbench/contrib/chat/common/chatVariables.ts
@@ -4,12 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { CancellationToken } from 'vs/base/common/cancellation';
-import { onUnexpectedExternalError } from 'vs/base/common/errors';
-import { Iterable } from 'vs/base/common/iterator';
-import { IDisposable, toDisposable } from 'vs/base/common/lifecycle';
+import { IDisposable } from 'vs/base/common/lifecycle';
+import { URI } from 'vs/base/common/uri';
+import { IRange } from 'vs/editor/common/core/range';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IChatModel } from 'vs/workbench/contrib/chat/common/chatModel';
-import { ChatRequestDynamicReferencePart, ChatRequestVariablePart, IParsedChatRequest } from 'vs/workbench/contrib/chat/common/chatParserTypes';
+import { IParsedChatRequest } from 'vs/workbench/contrib/chat/common/chatParserTypes';
 
 export interface IChatVariableData {
 	name: string;
@@ -36,6 +36,7 @@ export interface IChatVariablesService {
 	registerVariable(data: IChatVariableData, resolver: IChatVariableResolver): IDisposable;
 	hasVariable(name: string): boolean;
 	getVariables(): Iterable<Readonly<IChatVariableData>>;
+	getDynamicReferences(sessionId: string): ReadonlyArray<IDynamicReference>; // should be its own service?
 
 	/**
 	 * Resolves all variables that occur in `prompt`
@@ -43,77 +44,14 @@ export interface IChatVariablesService {
 	resolveVariables(prompt: IParsedChatRequest, model: IChatModel, token: CancellationToken): Promise<IChatVariableResolveResult>;
 }
 
-interface IChatData {
-	data: IChatVariableData;
-	resolver: IChatVariableResolver;
-}
-
-interface IChatVariableResolveResult {
+export interface IChatVariableResolveResult {
 	variables: Record<string, IChatRequestVariableValue[]>;
 	prompt: string;
 }
 
-export class ChatVariablesService implements IChatVariablesService {
-	declare _serviceBrand: undefined;
-
-	private _resolver = new Map<string, IChatData>();
-
-	constructor() {
-	}
-
-	async resolveVariables(prompt: IParsedChatRequest, model: IChatModel, token: CancellationToken): Promise<IChatVariableResolveResult> {
-		const resolvedVariables: Record<string, IChatRequestVariableValue[]> = {};
-		const jobs: Promise<any>[] = [];
-
-		const parsedPrompt: string[] = [];
-		prompt.parts
-			.forEach((part, i) => {
-				if (part instanceof ChatRequestVariablePart) {
-					const data = this._resolver.get(part.variableName.toLowerCase());
-					if (data) {
-						jobs.push(data.resolver(prompt.text, part.variableArg, model, token).then(value => {
-							if (value) {
-								resolvedVariables[part.variableName] = value;
-								parsedPrompt[i] = `[${part.text}](values:${part.variableName})`;
-							} else {
-								parsedPrompt[i] = part.text;
-							}
-						}).catch(onUnexpectedExternalError));
-					}
-				} else if (part instanceof ChatRequestDynamicReferencePart) {
-					// Maybe the dynamic reference should include a full IChatRequestVariableValue[] at the time it is inserted?
-					resolvedVariables[part.referenceText] = [{ level: 'full', value: part.data.toString() }];
-					parsedPrompt[i] = `[${part.text}](values:${part.referenceText})`;
-				} else {
-					parsedPrompt[i] = part.text;
-				}
-			});
-
-		await Promise.allSettled(jobs);
-
-		return {
-			variables: resolvedVariables,
-			prompt: parsedPrompt.join('')
-		};
-	}
-
-	hasVariable(name: string): boolean {
-		return this._resolver.has(name.toLowerCase());
-	}
-
-	getVariables(): Iterable<Readonly<IChatVariableData>> {
-		const all = Iterable.map(this._resolver.values(), data => data.data);
-		return Iterable.filter(all, data => !data.hidden);
-	}
-
-	registerVariable(data: IChatVariableData, resolver: IChatVariableResolver): IDisposable {
-		const key = data.name.toLowerCase();
-		if (this._resolver.has(key)) {
-			throw new Error(`A chat variable with the name '${data.name}' already exists.`);
-		}
-		this._resolver.set(key, { data, resolver });
-		return toDisposable(() => {
-			this._resolver.delete(key);
-		});
-	}
+export interface IDynamicReference {
+	range: IRange;
+	// data: any; // File details for a file, something else for a different type of thing, is it typed?
+	data: URI;
 }
+
